customize the user-agent of the OLTP exporter
Instead of the Go autoinst agent appearing to telemetry data receivers
as only the version of grpc-go against which it was built, this adds an
identifier to the User-Agent header via the dial options parameter
support within otlptracegrpc.

To help troubleshooting data transmission, runtime operating system and
CPU architecture is included in the user-agent within a parenthetical
comment in accordance with the header's specification.[1]

This replaces the low-level build up of a gRPC client connection with a
reuse of the OTel Go SDK's grpc trace client constructor. A happy
byproduct of using the SDK's constructor is that all of the
OTEL_EXPORTER_OTLP_* environment variables related to exporting traces
ought to be supported now, obviating the need to maintain code here to
retrieve, interpret, and handle them.

[1] RFC 7231 https://www.rfc-editor.org/rfc/rfc7231#section-5.5.3
    though more readable documentation about this header spec is
    available at MDN Web Docs:
    https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/User-Agent