Remove the shared vol from examples (#496)
This shared volumes is not needed. The application data does not need to
be shared between the auto-instrumentation image and the application.

Additionally, by using this shared volume it introduces a frustrating
bug. When the main app is changed, but the volume already exists the
existing volume will overwrite the change. Meaning the old app will
continue to run.