Switch `SpanEvent` times to use `time.Time` (#1101)
* Switch SpanEvent times to use time.Time

Use `time.Time`, a timestamp, to identify the start and end time of the
`SpanEvent` instead of representing them as nanoseconds since boot time
in a built-in `int64`.

This is motivate by the need to support things that report timestamps
for start/end times directly. Things like the custom SDK (see #1045).

To support the conversion for probes, the `utils.BootOffsetToTime`
function is added. This converts between the measured nanoseconds since
boot-time that an eBPF program measures, and a timestamp.

* Update internal/pkg/instrumentation/utils/kernel.go

Co-authored-by: Ron Federman <73110295+RonFed@users.noreply.github.com>

---------

Co-authored-by: Ron Federman <73110295+RonFed@users.noreply.github.com>