Test SDK TracerProvider
Move the tracerProvider type and related functionality to its own file
to help discoverability.

Use a singleton instance of the tracerProvider for
tracerProviderInstance instead of having a new value created each call
to TracerProvider.

Test the singleton nature of TracerProvider.

Test the concurrent safe behavior of the provided TracerProvider (part
of #1297).