Update `StructFieldConst` to find missing offsets at runtime (#1593)
* Add DWARF data processing to the process pkg

* Use process.DWARF in inspect app

* Add TargetDetails.OpenExe

Unify the opening of the target process exe file as a field.

* Support offset inspection with the inject pkg

* Update StructFieldConst to find missing offsets

If the offset index does not contain an entry for the desired struct
field, use the target process DWARF data to determine the offset.

* Update autosdk e2e test

Used to temporarily test the functionality. Should be reverted before
merge.

* Add changelog entry

* Revert "Update autosdk e2e test"

This reverts commit 0c18ff380655aef8447b6ddd5e4254285b10ac3f.

* Use full StructField in err msg

* Cached the DWARF data from the TargetDetails

* Doc TargetDetails.OpenExe

* Revert "Cached the DWARF data from the TargetDetails"

This reverts commit c5cb54b27b1c3d91bc0d8e2b74a76b7a1ff0c2fa.

Analysis has show that this caching could cause large allocations of
memory to be held for the lifetime of the Instrumentation. This is
removed to avoid that.