Move probe initialization to auto (#1906)
Instead of initializing all probes within the manger, have
Instrumentation handle the creation of all wanted probes.

This cleans up unneeded context being passed to the NewManager call
(i.e. globalImpl and Version()) and these values can be used where they
are defined.

This also is a first step in accepting probes as options to
Instrumentation which is a goal of #1105.