Lazy allocate memory for process  (#1945)
* Lazy allocate memory for process

Only allocate memory for a process if a probe requires the allocation.

Move the allocation functionality to be a method of the process.Info.
This is intrinsically linked to a single process, therefore, encapsulate
the functionality as a method.

* Flatten onceResult into Info

Based on feedback, since this is the only use of this type, do not
generalize it.

* Add tests for Info.Alloc

* Assert memory allocated after probe load