Split the integration-test into a matrix test (#2201)
* Split the integration-test into a matrix test

The log integration test is difficult to read. All integration tests are
sequentially in the same log message. This splits each e2e test into its
own job. With this the integration test that fails becomes immediately
apparent and job logs are specific to that test.

This does add overhead to the CI in that is spawns more jobs. That
should be considered when reviewing.

* Fix the cache keys