Replace BATS in kafka-go end-to-end test (#2217)
Part of a resolution for #2137.

This is a one-to-one replacement of the BATS testing as much as possible
with the one exception being that all producer spans are compared to the
consumer spans' parent span ID. This way when there is an ordering
difference and the parent is not the first span we do not introduce a
flaky test (like we have seen).