Handle "(devel)" version modules (#2437)
* Handle "(devel)" ver mods

Instead of failing to parse user modules with a version of "(devel)",
use a placeholder version value.

These modules are not known to be used by our probes. Failing to parse
their version and then failing the instrumentation is not desired
behavior. Instead, communicate the development state of the modules by
using a placeholder version.

* Add a changelog entry